+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Ignore errors when importing activities from outbox.
 - Make activity limit in outbox fetcher adjustable.
 - Changed `reset-subscriptions` command arguments (removes subscription options by default).
+- Return error if specified Monero account doesn't exist.
 
 ## [1.21.0] - 2023-04-12